swars: Routed some funcs through C, defined some vars
In short, preparing for rewrite of draw_new_panel().

+++ src/thing.h
@@ -321,7 +321,7 @@ struct TngUPerson
   short TempWeapon;
   short Stamina;
   short MaxStamina;
-  int WeaponsCarried;
+  ulong WeaponsCarried;
 };
 
 /** Structure for storing State of any Thing.
